# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Studio Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
  message: string;
}

export type UserPlan = (typeof UserPlan)[keyof typeof UserPlan];

export const UserPlan = {
  free: "free",
  pro: "pro",
  enterprise: "enterprise",
  vip: "vip",
  admin: "admin",
} as const;

export interface User {
  id: string;
  username: string;
  email?: string;
  plan: UserPlan;
  isAdmin: boolean;
  isVip: boolean;
  projectCount: number;
  buildsThisMonth: number;
  createdAt: string;
}

export type UpdateUserRequestPlan =
  (typeof UpdateUserRequestPlan)[keyof typeof UpdateUserRequestPlan];

export const UpdateUserRequestPlan = {
  free: "free",
  pro: "pro",
  enterprise: "enterprise",
  vip: "vip",
} as const;

export interface UpdateUserRequest {
  plan?: UpdateUserRequestPlan;
  isAdmin?: boolean;
}

export type ProjectType = (typeof ProjectType)[keyof typeof ProjectType];

export const ProjectType = {
  website: "website",
  mobile_app: "mobile_app",
  saas: "saas",
  automation: "automation",
  ai_tool: "ai_tool",
  game: "game",
} as const;

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  building: "building",
  ready: "ready",
  deployed: "deployed",
  failed: "failed",
} as const;

export interface Project {
  id: string;
  name: string;
  description?: string;
  type: ProjectType;
  status: ProjectStatus;
  prompt: string;
  framework?: string;
  gameEngine?: string;
  userId: string;
  thumbnailUrl?: string;
  deployedUrl?: string;
  createdAt: string;
  updatedAt: string;
  agentLogs: string[];
}

export type CreateProjectRequestType =
  (typeof CreateProjectRequestType)[keyof typeof CreateProjectRequestType];

export const CreateProjectRequestType = {
  website: "website",
  mobile_app: "mobile_app",
  saas: "saas",
  automation: "automation",
  ai_tool: "ai_tool",
  game: "game",
} as const;

export interface CreateProjectRequest {
  prompt: string;
  type: CreateProjectRequestType;
  name: string;
}

export type BuildLogLevel = (typeof BuildLogLevel)[keyof typeof BuildLogLevel];

export const BuildLogLevel = {
  info: "info",
  warn: "warn",
  error: "error",
  success: "success",
} as const;

export interface BuildLog {
  id: string;
  projectId: string;
  level: BuildLogLevel;
  message: string;
  agentName: string;
  timestamp: string;
}

export type FileNodeType = (typeof FileNodeType)[keyof typeof FileNodeType];

export const FileNodeType = {
  file: "file",
  directory: "directory",
} as const;

export interface FileNode {
  name: string;
  path: string;
  type: FileNodeType;
  content?: string;
  children?: FileNode[];
}

export type AgentCategory = (typeof AgentCategory)[keyof typeof AgentCategory];

export const AgentCategory = {
  software: "software",
  game_studio: "game_studio",
  business: "business",
  devops: "devops",
  design: "design",
  security: "security",
  database: "database",
} as const;

export type AgentStatus = (typeof AgentStatus)[keyof typeof AgentStatus];

export const AgentStatus = {
  idle: "idle",
  running: "running",
  completed: "completed",
  error: "error",
} as const;

export interface Agent {
  id: string;
  name: string;
  description: string;
  category: AgentCategory;
  status: AgentStatus;
  capabilities: string[];
  icon: string;
  color: string;
}

export type AgentRunRequestParameters = { [key: string]: unknown };

export interface AgentRunRequest {
  projectId?: string;
  task: string;
  parameters?: AgentRunRequestParameters;
}

export interface AgentRunResult {
  success: boolean;
  output: string;
  agentId: string;
  duration: number;
}

export type BuildStatus = (typeof BuildStatus)[keyof typeof BuildStatus];

export const BuildStatus = {
  queued: "queued",
  building: "building",
  success: "success",
  failed: "failed",
} as const;

export interface Build {
  id: string;
  projectId: string;
  status: BuildStatus;
  startedAt: string;
  completedAt?: string;
  logs: string[];
  deployedUrl?: string;
}

export type MarketplaceListingCategory =
  (typeof MarketplaceListingCategory)[keyof typeof MarketplaceListingCategory];

export const MarketplaceListingCategory = {
  app: "app",
  saas: "saas",
  game_template: "game_template",
  ai_tool: "ai_tool",
} as const;

export interface MarketplaceListing {
  id: string;
  projectId: string;
  title: string;
  description: string;
  category: MarketplaceListingCategory;
  price: number;
  isFree: boolean;
  sellerId: string;
  sellerName: string;
  downloads: number;
  rating: number;
  thumbnailUrl?: string;
  tags: string[];
  createdAt: string;
}

export type CreateListingRequestCategory =
  (typeof CreateListingRequestCategory)[keyof typeof CreateListingRequestCategory];

export const CreateListingRequestCategory = {
  app: "app",
  saas: "saas",
  game_template: "game_template",
  ai_tool: "ai_tool",
} as const;

export interface CreateListingRequest {
  projectId: string;
  title: string;
  description: string;
  category: CreateListingRequestCategory;
  price: number;
  tags: string[];
}

export type PlanName = (typeof PlanName)[keyof typeof PlanName];

export const PlanName = {
  free: "free",
  pro: "pro",
  enterprise: "enterprise",
  vip: "vip",
} as const;

export type PlanBillingPeriod =
  (typeof PlanBillingPeriod)[keyof typeof PlanBillingPeriod];

export const PlanBillingPeriod = {
  monthly: "monthly",
  yearly: "yearly",
} as const;

export interface PlanLimits {
  buildsPerMonth: number;
  deployments: number;
  aiUsageTokens: number;
  projects: number;
  teamMembers: number;
}

export interface Plan {
  id: string;
  name: PlanName;
  displayName: string;
  price: number;
  billingPeriod: PlanBillingPeriod;
  features: string[];
  limits: PlanLimits;
}

export type AnalyticsOverviewRevenueByPlan = { [key: string]: number };

export interface TimeSeriesPoint {
  date: string;
  value: number;
}

export interface AnalyticsOverview {
  totalUsers: number;
  totalProjects: number;
  totalBuilds: number;
  totalRevenue: number;
  activeAgents: number;
  marketplaceListings: number;
  buildsToday: number;
  newUsersThisWeek: number;
  revenueByPlan: AnalyticsOverviewRevenueByPlan;
  buildsOverTime: TimeSeriesPoint[];
}

export type UserAnalyticsProjectsByType = { [key: string]: number };

export interface UserAnalytics {
  projectsCreated: number;
  buildsThisMonth: number;
  buildsRemaining: number;
  totalDeployments: number;
  aiTokensUsed: number;
  aiTokensRemaining: number;
  projectsByType: UserAnalyticsProjectsByType;
  activityTimeline: TimeSeriesPoint[];
}

export type ListProjectsParams = {
  type?: ListProjectsType;
  status?: ListProjectsStatus;
};

export type ListProjectsType =
  (typeof ListProjectsType)[keyof typeof ListProjectsType];

export const ListProjectsType = {
  website: "website",
  mobile_app: "mobile_app",
  saas: "saas",
  automation: "automation",
  ai_tool: "ai_tool",
  game: "game",
  all: "all",
} as const;

export type ListProjectsStatus =
  (typeof ListProjectsStatus)[keyof typeof ListProjectsStatus];

export const ListProjectsStatus = {
  building: "building",
  ready: "ready",
  deployed: "deployed",
  failed: "failed",
  all: "all",
} as const;

export type ListBuildsParams = {
  projectId?: string;
};

export type ListMarketplaceListingsParams = {
  category?: ListMarketplaceListingsCategory;
  search?: string;
};

export type ListMarketplaceListingsCategory =
  (typeof ListMarketplaceListingsCategory)[keyof typeof ListMarketplaceListingsCategory];

export const ListMarketplaceListingsCategory = {
  app: "app",
  saas: "saas",
  game_template: "game_template",
  ai_tool: "ai_tool",
  all: "all",
} as const;

export type ListUsersParams = {
  plan?: string;
};