# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Studio Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects for user
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [website, mobile_app, saas, automation, ai_tool, game, all]
        - name: status
          in: query
          schema:
            type: string
            enum: [building, ready, deployed, failed, all]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project from prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /projects/{id}/build-logs:
    get:
      operationId: getProjectBuildLogs
      tags: [projects]
      summary: Get build logs for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Build logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BuildLog"
  /projects/{id}/files:
    get:
      operationId: getProjectFiles
      tags: [projects]
      summary: Get file tree for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileNode"
...
  /marketplace/listings:
    get:
      operationId: listMarketplaceListings
      tags: [marketplace]
      summary: List marketplace items
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [app, saas, game_template, ai_tool, all]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Marketplace listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceListing"
    post:
      operationId: createMarketplaceListing
      tags: [marketplace]
      summary: Publish project to marketplace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingRequest"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceListing"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [website, mobile_app, saas, automation, ai_tool, game]
        status:
          type: string
          enum: [building, ready, deployed, failed]
        prompt:
          type: string
        framework:
          type: string
        gameEngine:
          type: string
        userId:
          type: string
        thumbnailUrl:
          type: string
        deployedUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        agentLogs:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - type
        - status
        - prompt
        - userId
        - createdAt
        - updatedAt
        - agentLogs
    CreateProjectRequest:
      type: object
      properties:
        prompt:
          type: string
        type:
          type: string
          enum: [website, mobile_app, saas, automation, ai_tool, game]
        name:
          type: string
      required:
        - prompt
        - type
        - name
    BuildLog:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        level:
          type: string
          enum: [info, warn, error, success]
        message:
          type: string
        agentName:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - level
        - message
        - agentName
        - timestamp
    FileNode:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        content:
          type: string
        children:
          type: array
          items:
            $ref: "#/components/schemas/FileNode"
      required:
        - name
        - path
        - type
...
    MarketplaceListing:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [app, saas, game_template, ai_tool]
        price:
          type: number
        isFree:
          type: boolean
        sellerId:
          type: string
        sellerName:
          type: string
        downloads:
          type: integer
        rating:
          type: number
        thumbnailUrl:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - description
        - category
        - price
        - isFree
        - sellerId
        - sellerName
        - downloads
        - rating
        - tags
        - createdAt
    CreateListingRequest:
      type: object
      properties:
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [app, saas, game_template, ai_tool]
        price:
          type: number
        tags:
          type: array
          items:
            type: string
      required:
        - projectId
        - title
        - description
        - category
        - price
        - tags
...