# lib/api-zod/src/generated/types/buildLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Studio Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BuildLogLevel } from "./buildLogLevel";

export interface BuildLog {
  id: string;
  projectId: string;
  level: BuildLogLevel;
  message: string;
  agentName: string;
  timestamp: Date;
}